\name{print.GFA}
\alias{print.GFA}
\title{
Print
}
\description{
Print the summarized results of the estimated group factor model, such as the estimated global and local factors.
}
\usage{
\method{print}{GFA}(x, ...)
}
\arguments{
  \item{x}{the \code{GFA} object returned from the algorithm.}
  \item{\dots}{additional print arguments.}

}
\value{
No return value, called for side effects
}
