% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{print.GFD}
\alias{print.GFD}
\title{Print Method for GFD Objects}
\usage{
\method{print}{GFD}(x, ...)
}
\arguments{
\item{x}{The \code{GFD} object returned from \code{\link{gendata}}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
No return value, called for side effects.
}
\description{
Print the summary of the generated grouped data.
}
