\name{TraceRatio}
\alias{TraceRatio}
\title{
Trace ratios
}
\description{
Evaluation of the estimated factors by trace ratios, the values is between 0 and 1, higher values means better estimation.
}
\usage{
TraceRatio(G, Ghat)
}
\arguments{
  \item{G}{the true factors.}
  \item{Ghat}{the estimated factors.}
}
\value{
  \item{ratios}{the trace ratios: TR1, TR2.}
}

\author{
Jiaqi Hu
}
