% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{print.GFA}
\alias{print.GFA}
\title{Print Method for GFA Objects}
\usage{
\method{print}{GFA}(x, ...)
}
\arguments{
\item{x}{The \code{GFA} object returned from the estimation algorithms (e.g., \code{APM}, \code{CCA}, \code{GCC}, \code{CP}).}

\item{...}{Additional arguments passed to methods.}
}
\value{
No return value, called for side effects.
}
\description{
Print the summarized results of the estimated group factor model, such as the estimated global and local factor numbers and reference statistics.
}
